+++ service/featuregate/gates.go
@@ -46,30 +46,24 @@ type Registry struct {
 
 // Apply a configuration in the form of a map of Gate identifiers to boolean values.
 // Sets only those values provided in the map, other gate values are not changed.
-// Deprecated: [v0.56.0] Use MustApply instead.
-func (r *Registry) Apply(cfg map[string]bool) {
+func (r *Registry) Apply(cfg map[string]bool) error {
 	r.mu.Lock()
 	defer r.mu.Unlock()
 	for id, val := range cfg {
-		if g, ok := r.gates[id]; ok {
-			g.Enabled = val
-			r.gates[g.ID] = g
+		g, ok := r.gates[id]
+		if !ok {
+			return fmt.Errorf("feature gate %s is unregistered", id)
 		}
+		g.Enabled = val
+		r.gates[g.ID] = g
 	}
+	return nil
 }
 
-// MustApply a configuration in the form of a map of Gate identifiers to boolean values.
-// Sets only those values provided in the map, other gate values are not changed.
+// Deprecated: [v0.58.0] Use Apply instead.
 func (r *Registry) MustApply(cfg map[string]bool) {
-	r.mu.Lock()
-	defer r.mu.Unlock()
-	for id, val := range cfg {
-		if g, ok := r.gates[id]; ok {
-			g.Enabled = val
-			r.gates[g.ID] = g
-		} else {
-			panic(fmt.Sprintf("feature gate %s is unregistered", id))
-		}
+	if err := r.Apply(cfg); err != nil {
+		panic(err)
 	}
 }